Fix launching the game, spotless
