+++ dependencies.gradle
@@ -58,7 +58,7 @@ dependencies {
     compileOnly("TGregworks:TGregworks:1.7.10-GTNH-1.0.26:deobf") {transitive = false}
     compileOnly("com.github.GTNewHorizons:EnderCore:0.4.6:dev") { transitive = false }
     implementation("com.github.GTNewHorizons:Galacticraft:3.1.5-GTNH:dev") { transitive = false }
-    implementation("com.github.GTNewHorizons:TinkersConstruct:1.12.1-GTNH:dev") { transitive = false }
+    implementation("com.github.GTNewHorizons:TinkersConstruct:1.12.1-GTNH:dev")
     compileOnly("com.github.GTNewHorizons:Chisel:2.15.0-GTNH:dev") { transitive = false }
     compileOnly("com.github.GTNewHorizons:Translocators:1.2.1:dev") { transitive = false }
     compileOnly("curse.maven:cofh-core-69162:2388751") { transitive = false }
@@ -69,7 +69,7 @@ dependencies {
     compileOnly('com.github.GTNewHorizons:HoloInventory:2.4.10-GTNH:dev') { transitive = false }
     compileOnly("curse.maven:extra-utilities-225561:2264384") {transitive=false}
     compileOnly('curse.maven:minefactory-reloaded-66672:2366150') { transitive = false }
-    compileOnly("com.github.GTNewHorizons:OpenComputers:1.10.11-GTNH:api") {transitive = false}
+    compileOnly("com.github.GTNewHorizons:OpenComputers:1.10.11-GTNH:dev") {transitive = false}
     // https://www.curseforge.com/minecraft/mc-mods/advancedsolarpanels
     implementation('curse.maven:advsolar-362768:2885953')
