+++ src/main/java/gtPlusPlus/xmod/thaumcraft/commands/CommandDumpAspects.java
@@ -30,7 +30,11 @@ public class CommandDumpAspects implements ICommand {
 
     @Override
     public int compareTo(final Object o) {
-        return 0;
+        if (o instanceof ICommand c) {
+            return this.getCommandName()
+                .compareTo(c.getCommandName());
+        }
+        return -1;
     }
 
     @Override
@@ -77,7 +81,7 @@ public class CommandDumpAspects implements ICommand {
     }
 
     @Override
-    public List<?> addTabCompletionOptions(final ICommandSender var1, final String[] var2) {
+    public List<String> addTabCompletionOptions(final ICommandSender var1, final String[] var2) {
         return null;
     }
